refactor(core): move tool-execution guards out of the Gemini module
`guardToolExecutor` and its guards type were written for the native Gemini
loops and named accordingly. They are no longer Gemini-specific: both
Vertex+Claude loops call `guardToolExecutor` directly — four call sites in
googleVertex/client.ts, inside the Anthropic stream and generate turns — so the
Anthropic paths were importing a tool-execution primitive out of
`providers/googleNativeGemini3/utils.ts` behind a type called
`GeminiToolExecutionGuards`. Accurate about where the code was written,
misleading about what depends on it, and the kind of thing that sends the next
reader to the wrong provider family.

`GeminiToolExecutionGuards` is now `ToolExecutionGuards`, and the function moves
to `core/toolExecutionGuards.ts`. Nothing in it is provider-specific: it takes
an executor plus a guards object and returns a wrapped executor, built only on
`raceWithAbort` and `withTimeout`.

`buildDedupedEngineTools` deliberately STAYS in the Gemini module. It takes
`NativeToolDeclarationsResult` and is only called from the two Gemini providers,
so it is correctly located — the point here is to separate the shared primitive
from the Gemini-shaped wrapper around it, not to relocate everything the file
happens to hold.

Behaviour is unchanged, and that was checked rather than assumed. Call sites
before and after:

  googleVertex/client.ts            4 -> 4   (all four Anthropic sites intact)
  googleNativeGemini3/utils.ts      2 -> 1   (the definition left; its one call
                                              inside buildDedupedEngineTools
                                              stays)
  core/geminiLoopAdapter.ts         1 -> 1
  total invocations                 6 -> 6

Verified: typecheck 4816 files 0 errors, eslint 0 errors (56 warnings, the same
pre-existing set release carries), prettier clean, and the four loop
characterization suites unchanged at vertex-claude 7, anthropic-loop 6,
vertex-loop 14, aistudio-loop 10, plus providers-mocked 54/54.
