+++ docs/api/README.md
@@ -1303,7 +1303,8 @@ console.log(result.content);
 - [AgenticLoopToolFailureBreaker](type-aliases/AgenticLoopToolFailureBreaker.md)
 - [AgenticLoopAdapter](type-aliases/AgenticLoopAdapter.md)
 - [AnthropicLoopAdapterConfig](type-aliases/AnthropicLoopAdapterConfig.md)
-- [GeminiToolExecutionGuards](type-aliases/GeminiToolExecutionGuards.md)
+- [ToolExecutionGuards](type-aliases/ToolExecutionGuards.md)
+- [~~GeminiToolExecutionGuards~~](type-aliases/GeminiToolExecutionGuards.md)
 - [GeminiStepRaw](type-aliases/GeminiStepRaw.md)
 - [GeminiTurnContent](type-aliases/GeminiTurnContent.md)
 - [GeminiLoopAdapterCoreConfig](type-aliases/GeminiLoopAdapterCoreConfig.md)
